+++ package.json
@@ -45,6 +45,7 @@
     "next": "12.2.0",
     "next-auth": "^4.20.1",
     "next-themes": "^0.2.1",
+    "next-use-contextual-routing": "^3.0.1",
     "nodemailer": "^6.7.5",
     "nprogress": "^0.2.0",
     "react": "18.2.0",
