+++ src/components/CommentActionModal.tsx
@@ -1,6 +1,5 @@
 import { CommentWithChildren } from "@utils/types";
 import { Dispatch, SetStateAction } from "react";
-import { MdClose } from "react-icons/md";
 import Comment from "./Comment";
 import CommentField from "./CommentField";
 import EditCommentForm from "./EditCommentForm";
@@ -13,15 +12,6 @@ type Props = {
   openState: [boolean, Dispatch<SetStateAction<boolean>>];
 };
 
-const CloseButton = ({ onClick }: { onClick: () => void }) => (
-  <button
-    onClick={onClick}
-    className="absolute z-[3] rounded-full top-0 right-0 p-2 bg-emerald-500 dark:bg-teal-900"
-  >
-    <MdClose className="w-5 h-5 text-white" />
-  </button>
-);
-
 const CommentActionModal: React.FC<Props> = ({
   openState,
   parentComment,
@@ -32,10 +22,8 @@ const CommentActionModal: React.FC<Props> = ({
   const closeModal = () => setOpen(false);
 
   return (
-    <Modal openState={openState} alwaysCentered>
+    <Modal openState={openState} alwaysCentered closeButton>
       <div className="relative overflow-hidden bg-white dark:bg-neutral-900 rounded-lg px-4 pt-5 pb-4 shadow-xl sm:my-8 sm:w-full sm:max-w-lg sm:p-6">
-        <CloseButton onClick={closeModal} />
-
         <Comment comment={parentComment} compact hideReplies hideActions />
 
         <div className="w-full mt-5">
