+++ src/components/Modal.tsx
@@ -1,16 +1,42 @@
 import { Fragment, Dispatch, MutableRefObject, SetStateAction } from "react";
 import { Dialog, Transition } from "@headlessui/react";
+import { MdClose } from "react-icons/md";
+import ShouldRender from "./ShouldRender";
+
+const CloseButton = ({ onClick }: { onClick: () => void }) => (
+  <button
+    onClick={onClick}
+    className="absolute z-[3] rounded-full top-0 right-0 p-2 bg-emerald-500 dark:bg-teal-900"
+  >
+    <MdClose className="w-5 h-5 text-white" />
+  </button>
+);
 
 export const Modal: React.FC<{
   openState: [boolean, Dispatch<SetStateAction<boolean>>];
   initialFocus?: MutableRefObject<null>;
   children: React.ReactNode;
   alwaysCentered?: boolean;
+  onClose?: () => void;
+  closeButton?: boolean;
 }> & {
   Title: typeof Dialog.Title;
   Description: typeof Dialog.Description;
-} = ({ openState, initialFocus, alwaysCentered = false, children }) => {
+} = ({
+  openState,
+  initialFocus,
+  alwaysCentered = false,
+  children,
+  onClose,
+  closeButton,
+}) => {
   const [open, setOpen] = openState;
+
+  const handleClose = () => {
+    if (onClose) onClose();
+
+    setOpen(false);
+  };
 
   return (
     <Transition.Root show={open} as={Fragment}>
@@ -18,7 +44,7 @@ export const Modal: React.FC<{
         as="div"
         className="fixed inset-0 z-[100] overflow-y-auto"
         initialFocus={initialFocus}
-        onClose={setOpen}
+        onClose={handleClose}
       >
         <div
           className={`flex min-h-screen ${
@@ -54,6 +80,9 @@ export const Modal: React.FC<{
             leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
           >
             <div className="relative inline-block text-left align-bottom sm:align-middle">
+              <ShouldRender if={closeButton}>
+                <CloseButton onClick={handleClose} />
+              </ShouldRender>
               {children}
             </div>
           </Transition.Child>
